refactor test code to allocate new handler and reuse test code
