+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/admin_appeal_withdrawn/AdminAppealWithdrawnDwpStateToWithdrawnHandler.java
@@ -1,0 +1,19 @@
+package uk.gov.hmcts.reform.sscs.ccd.presubmit.admin_appeal_withdrawn;
+
+import uk.gov.hmcts.reform.sscs.ccd.callback.Callback;
+import uk.gov.hmcts.reform.sscs.ccd.callback.CallbackType;
+import uk.gov.hmcts.reform.sscs.ccd.callback.PreSubmitCallbackResponse;
+import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
+import uk.gov.hmcts.reform.sscs.ccd.presubmit.PreSubmitCallbackHandler;
+
+public class AdminAppealWithdrawnDwpStateToWithdrawnHandler implements PreSubmitCallbackHandler<SscsCaseData> {
+    @Override
+    public boolean canHandle(CallbackType callbackType, Callback<SscsCaseData> callback) {
+        return false;
+    }
+
+    @Override
+    public PreSubmitCallbackResponse<SscsCaseData> handle(CallbackType callbackType, Callback<SscsCaseData> callback) {
+        return null;
+    }
+}
