+++ src/test/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/admin_appeal_withdrawn/AdminAppealWithdrawnBase.java
@@ -1,0 +1,52 @@
+package uk.gov.hmcts.reform.sscs.ccd.presubmit.admin_appeal_withdrawn;
+
+import static com.fasterxml.jackson.databind.DeserializationFeature.READ_ENUMS_USING_TO_STRING;
+import static com.fasterxml.jackson.databind.DeserializationFeature.READ_UNKNOWN_ENUM_VALUES_USING_DEFAULT_VALUE;
+import static com.fasterxml.jackson.databind.SerializationFeature.WRITE_ENUMS_USING_TO_STRING;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import java.io.File;
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.util.Objects;
+import org.apache.commons.io.FileUtils;
+import org.junit.Before;
+import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
+import uk.gov.hmcts.reform.sscs.ccd.callback.Callback;
+import uk.gov.hmcts.reform.sscs.ccd.deserialisation.SscsCaseCallbackDeserializer;
+import uk.gov.hmcts.reform.sscs.ccd.domain.EventType;
+import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
+
+public class AdminAppealWithdrawnBase {
+    private ObjectMapper mapper;
+
+    @Before
+    public void setUp() {
+        Jackson2ObjectMapperBuilder objectMapperBuilder =
+            new Jackson2ObjectMapperBuilder()
+                .featuresToEnable(READ_ENUMS_USING_TO_STRING)
+                .featuresToEnable(READ_UNKNOWN_ENUM_VALUES_USING_DEFAULT_VALUE)
+                .featuresToEnable(WRITE_ENUMS_USING_TO_STRING)
+                .serializationInclusion(JsonInclude.Include.NON_ABSENT);
+
+        mapper = objectMapperBuilder.createXmlMapper(false).build();
+        mapper.registerModule(new JavaTimeModule());
+    }
+
+    Callback<SscsCaseData> buildTestCallback(EventType eventType) throws IOException {
+        if (eventType == null) {
+            return null;
+        }
+        String json = fetchData("callback/adminAppealWithdrawnCallback.json");
+        String jsonCallback = json.replace("EVENT_ID_PLACEHOLDER", eventType.getCcdType());
+        SscsCaseCallbackDeserializer sscsCaseCallbackDeserializer = new SscsCaseCallbackDeserializer(mapper);
+        return sscsCaseCallbackDeserializer.deserialize(jsonCallback);
+    }
+
+    String fetchData(String s) throws IOException {
+        String file = Objects.requireNonNull(getClass().getClassLoader().getResource(s)).getFile();
+        return FileUtils.readFileToString(new File(file), StandardCharsets.UTF_8.name());
+    }
+}
