+++ src/test/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/admin_appeal_withdrawn/AdminAppealWithdrawnDwpStateToWithdrawnHandlerTest.java
@@ -1,0 +1,16 @@
+package uk.gov.hmcts.reform.sscs.ccd.presubmit.admin_appeal_withdrawn;
+
+import org.junit.Test;
+
+public class AdminAppealWithdrawnDwpStateToWithdrawnHandlerTest extends AdminAppealWithdrawnBase {
+
+    @Test
+    public void canHandle() {
+        AdminAppealWithdrawnDwpStateToWithdrawnHandler handler = new AdminAppealWithdrawnDwpStateToWithdrawnHandler();
+//        handler.canHandle(CallbackType.ABOUT_TO_SUBMIT, )
+    }
+
+    @Test
+    public void handle() {
+    }
+}
